pyproject.toml
[tool.poetry]
name = "codstattracker"
version = "0.1.0"
description = "App to poll and collect your COD stats"
authors = ["Daniel Daniels <danields761@gmail.com>"]
license = "NPOSL-3.0"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "*"
requests = "^2.*"
pydantic = "^1.7"
mysql-connector-python = "^8.0"
SQLAlchemy = "^1.3"
pydantic-settings = "^0.2"
sentry-sdk = "^0.19"
yoyo-migrations = "^7.2.1"

[tool.poetry.dev-dependencies]
black = "*"
pytest = "^6.1.2"
isort = "^5.6.4"

[tool.poetry.scripts]
cst-poller = 'codstattracker.poller.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Tools configuration
[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 79
known_first_party = ["codstattracker", "tests"]
